paddleocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
paddlepaddle==3.0.0
paddleocr==2.8.1
python-multipart==0.0.12
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
